requirements.txt
pymongo[srv]
uvicorn
fastapi
paho-mqtt==1.5.1
